Extract to values package shared logic for merging configmaps and secrets
